Add test harness for front-end code, expand line-status.js and trip.js
We did not have a good way to write code to test browserified components,
so this commit adds the ability to test modules using the normal chai/mocha
workflow, without breaking the associations to the correct production
versions of lodash and backbone. It's a little convoluted, but it is in the
service of accurate environment reproduction.