Fix bug where avg wait disappears when returning to list from station
Because this can happen within 15 seconds, the refresh was never getting triggered and
therefore the status view was not getting any message to refresh. This both force-renders
the view, and also makes stepping through the app more efficient because we now persist
references to commonly needed status collections at the data store level within a client
session.