Re-implement the application router using StateMan
The primary purpose of this re-implementation is to make the processing
of each route, as well as its entry and exit, more obvious and less
dependent on Backbone (in preparation for a migration off of the same.)
Ampersand-Router was considered, but I opted to use StateMan (a) to
give myself an excuse to experiment with a standalone router that uses
a different paradigm than I'm used to, and (b) because I subjectively
felt it was a cleaner API.

We are not taking advantage of route nesting in StateMan, because
I don't fully grok nested routes and the application was implemented
assuming an un-nestable system.

We also avoid the use of StateMan's auto-link functionality, because
we use actual links in order to set ourselves up to have the app
be able to render itself on the backend down the road.

We are also doing one unusual thing with StateMan, which is that I do
not see the use-case for separating `enter` and `update`, so `update`
just proxies through to `enter` wherever it is called.

Benefits to the new approach:

- Gives us a reusable construct for setting the document.title
- Gives us a router with exit handlers we could use to tear down UI
- Gives us slightly cleaner 404 handling, by auto-404'ing some routes
  (e.g. /:line for values of :line other than red, green, orange)