Switch from Nunjucks to Combyne for client-side templating
[CombynExpress](https://github.com/tbranyen/combynexpress) and [Combynify](https://github.com/chesles/combynify) provide almost the same level of robust templating as Nunjucks does, but the Express and Browserify versions of Combyne feel less separate than they do when using `nunjucksify`, and I get that fuzzy locavore feeling because it's a @tbranyen project.

This PR installs visit-combyne as a dependency; visit-combyne is contained within combynify, but combynify will not work properly with less than visit-combyne 1.0.2 (due to a recursion issue that causes Combynify to miss several required filters), so the local installation is done to force the latest version.

This change also revealed some logical errors in both the templates and views themselves, as well as in the router, so this will also make 404/500 errors more predictable on the Express side.