Show a "Loading" indicator for predictions (fixes #22)
Per early feedback from @iros, it was worrying that the app will display
"No Trains" when it simply hasn't gotten the data back yet. Adding a
loading state while the request is in progress reduces concern that no
trains are in service and improves communication about the state of the app.